migrated collection routes

+++ apps/api/src/routes/collection.ts
@@ -1,36 +1,48 @@
 import { card, collectionCard } from "../db/schema";
-import { eq, and, sql, isNull } from "drizzle-orm";
+import { eq, sql, isNull } from "drizzle-orm";
 import { base } from "../orpc";
 import {
   memberOfLeague,
   selfOrLeagueOwner,
 } from "../middleware/leagueMembership";
+import { ORPCError } from "@orpc/server";
 
 const getCollection = base.collection.get
   .use(memberOfLeague)
   .handler(async ({ input, context }) => {
-    const cards = await context.env.db.query.collectionCard.findMany({
+    const res = await context.env.db.query.leagueMember.findFirst({
       where: {
         leagueId: input.leagueId,
         userId: input.userId,
       },
-      with: { card: true },
+      with: {
+        collection: { with: { cardQuantities: { with: { card: true } } } },
+      },
     });
 
-    return { cards };
+    if (!res) throw new ORPCError("NOT_FOUND");
+
+    return { cards: res.collection.cardQuantities };
   });
 
 const setCollection = base.collection.set
   .use(selfOrLeagueOwner)
   .handler(async ({ input, context, errors }) => {
     context.env.db.transaction((tx) => {
+      const res = tx.query.leagueMember
+        .findFirst({
+          columns: { collectionId: true },
+          where: {
+            userId: input.userId,
+            leagueId: input.leagueId,
+          },
+        })
+        .sync();
+      if (!res) throw new ORPCError("NOT_FOUND");
+      const { collectionId } = res;
+
       tx.delete(collectionCard)
-        .where(
-          and(
-            eq(collectionCard.leagueId, input.leagueId),
-            eq(collectionCard.userId, input.userId),
-          ),
-        )
+        .where(eq(collectionCard.collectionId, collectionId))
         .run();
 
       const valuesSql = sql.join(
@@ -63,16 +75,11 @@ const setCollection = base.collection.set
           input.cards.map(({ cardName, quantity }) => ({
             cardName,
             quantity,
-            leagueId: input.leagueId,
-            userId: input.userId,
+            collectionId,
           })),
         )
         .onConflictDoUpdate({
-          target: [
-            collectionCard.leagueId,
-            collectionCard.userId,
-            collectionCard.cardName,
-          ],
+          target: [collectionCard.collectionId, collectionCard.cardName],
           set: {
             quantity: sql`${collectionCard.quantity} + excluded.quantity`,
           },
